Add Zstandard compression option for append-optimized tables
This is a backport of
https://github.com/greenplum-db/gpdb/commit/724f9d27b160ec3f87a5ce95eecc605808be3ced.

This commit provides all necessary changes to implement Zstandard compression,
as well as regression test and changes in documentation.

Add a new compression option for append-optimized tables, "zstd". It is
generally faster than zlib or quicklz, and compresses better. Or at least
it can be faster or compress better, if not both at the same time, by
adjusting the compression level. A major advantage of Zstandard is the wide
range of tuning, to choose the trade-off between compression speed and
ratio.

Using the new option requires building the server with the `libzstd` library.
A new `--with-zstd` option is added for that. The default is to build without
libzstd, for now, but we should probably change the default to be on, after
we have had a chance to update all the buildfarm machines to have libzstd.

Original commit by Ivan Leskin, Dmitriy Pavlov, Anton Chevychalov and Heikki Linnakangas.