CRC32C optimization for POWER8
(in development)

Implement an optimized version of CRC32C calculation algorithm for IBM POWER8.

The algorithm is a modified version of 'https://github.com/antonblanchard/crc32-vpmsum'. It uses POWER8 vector intrinsics to optimize the computation. The description of the original algorithm says we can expect the optimized version to be 30 times faster than a conventional slice-by-8 algorithm (which is the current implementation of CRC32 for all POWER CPUs).