ADBDEV-36 Fixup module's compatibility checker
There is a protective code in every module and gpdb, that
compares versions inside module and gpdb. Due to legacy reasons
it compares up to second digit (5.4 != 5.2). The truth that
there is a real compatibility acros 5.x (5.x == 5.y for x,i in INT)

To achive that compatitibilty upstream stops to change version inside
gpdb (it's 5.0.0 always).

That commit changes checker logic and now it compares fist digit only.