Stop using strxfrm() for string sorting.
strxfrm() is buggy on some platforms and locales. We found this out the
hard way in the PostgreSQL community, when we tried to use strxfrm() for
string sorting in PostgreSQL 9.5. We've been doing it on Greenplum,
anyway, but it's just as buggy for Greenplum as it was for PostgreSQL 9.5.

Fixes https://github.com/greenplum-db/gpdb/issues/6586

Discussion: https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/MUplsaJHLg0

Co-Authored-By: Ivan Leskin leskin.in@arenadata.io
Co-Authored-By: Denis Smirnov sd@arenadata.io