Print block directory enty map in logs
This commit simply logs block directory entry map for every column
group in AO/AOC relation. We can see that AOC table has absolutely
different number of variable sized blocks (and the number of tuples
in each) for every column. But the total number of tuples for each
column with equal segment file number is the same among all columns.