cdbdisp_buildUtilityQueryParams: Replace NULL debug_query_string with ""
Closes https://github.com/greenplum-db/gpdb/issues/10918.

debug_query_string global variable is used by
cdbdisp_buildUtilityQueryParams() to pass the original query to segments
executing utility queries.

There is a case when debug_query_string is NULL (see the linked issue).

Handle this case by checking NULL-ness of the variable and changing it
to empty string if necessary.