Implement two-stage sampling for AO/AOC analyze
To calculate statistics at first we have to retrieve in reservoir some
representative sample from relation. PostgreSQL uses two stage sampling
to avoid sequential scan of the whole relation:

* at first it samples blocks with algorithm S from Knuth
* for each block it samples tuples with algorithm Z from Vitter

To make it work every AM has to know how to estimate block number for
any relation. These blocks should be of the fix size to make algorithm
S from Knuth work correctly. Previously, AO/AOC relations didn't use two
stage sampling and sequentially scanned the whole relation. The main
reason for it was the fact that segment files are constructed from variable
sized blocks (that are not suitable for algorithm S from Knuth).

Current implementation treats AO/AOC relations as a collection of segment
files with a global sequence of rows among them:

* segno 1 with 100 rows - its rows are treated as global positions 0-99
* segno 2 with 200 rows - its rows are treated as global positions 100-299

This global sequence of rows is divided on some arbitrary number of "logical"
analyze blocks fitting BlockNumber (uint32). These blocks are not
interconnected with variable sized physical blocks of segment files -
it is simply a range of global row positions. So block sampling by
algorithm S from Knuth gives us the starting global position of a target
block (the size of the block is fixed and known). We get a benefit of this
schema when the number of blocks is more that 300 * statistics target.
So if we are lucky enough there would be several physical variable
sized blocks between logical blocks and we can just read their headers
without decompressing data. This is the main idea of speeding up AO/AOC
analyze. After that we simply iterate among all rows in a logical analyze
block and fill reservoir with samples.

Current implementation is suboptimal in getting relation's analyze block count.
To always fit BlockNumber (uint32) we use a fix sized block of 65536 rows. But
as far as we should collect 300 * statistic target (default = 100) blocks
of 65536 rows, we get benefit from two stage sampling starting from ~2 billion
rows per GP instance. It is made just to simplify the code but in fact we
can use any logic to get block count for AO/AOC relation (not so straightforward
as we do now).

Another suboptimal thing is AOC projections - I didn't find a way to get columns
required for current analyze and project all of them at the moment.