Implement two-stage sampling for AO/AOC analyze
To calculate statistics at first we have to retrieve in reservoir some
representative sample from relation. PostgreSQL uses two stage sampling
to avoid sequential scan of the whole relation:

* at first it samples blocks with algorithm S from Knuth
* for each block it samples tuples with algorithm Z from Vitter

To make it work every AM has to know how to estimate block number for
any relation. These blocks should be of the fix size to make algorithm
S from Knuth work correctly. Previously, AO/AOC relations didn't use two
stage sampling and sequentially scanned the whole relation. The main
reason for it was the fact that segment files are constructed from variable
sized blocks (that are not suitable for algorithm S from Knuth).

Current implementation treats AO/AOC relations as a collection of segment
files with a global sequence of rows among them:

* segno 1 with 100 rows - its rows are treated as global positions 0-99
* segno 2 with 200 rows - its rows are treated as global positions 100-299

This global sequence of rows is divided on some arbitrary number of "logical"
analyze blocks fitting BlockNumber (uint32). These blocks are not
interconnected with variable sized physical blocks of segment files -
it is simply a range of global row positions. So block sampling by
algorithm S from Knuth gives us the starting global position of a target
block (the size of the block is fixed and known). We get a benefit of this
schema when the number of blocks is more that 300 * statistics target.
So if we are lucky enough there would be several physical variable
sized blocks between logical blocks and we can just read their headers
without decompressing data. This is the main idea of speeding up AO/AOC
analyze. After that we simply iterate among all rows in a logical analyze
block and fill reservoir with samples.

To divide AO/AOC table into logical blocks we introduce a new GUC
'gp_appendonly_analyze_block_size' controlling the default analyze
block size. We divide AO/AOC table with logical blocks of this size. If
the result number of blocks doesn't fit BlockNumber we replace it
with the maximum possible number of blocks (2^33-1).

Current implementation is suboptimal in AOC projections at the beginning
of analyze scan: AM API doesn't allow us to project only required columns
for analyze scan and we have to scan all of them. It would be nice to improve
AM cause a simple scan is initialised with such projection, but analyze scan is
not.

I would also like to explain why we use fix sized logical blocks sampling with
algorithm S from Knuth instead of weighted random sampling (WRS) algorithms
(A-Chao or Efraimidis and Spirakis). At first I tried them but was confused with
AOC sampling. Every column has its own variable sized blocks with different
sizes and I didn't know what column to choose for WRS. The only thing in common
among all columns segment files was the row count, so that was the reason
to chose sampling based on fix sized logical blocks in a global sequence of rows
in a table.