GPORCA: map DATE type values to LINT # of days instead of DOUBLE mcsecs
The commit 9b42761 had intoduced the change in computing of frequency
for singleton bucket produced from some enclosing one. Instead of
division frequency of enclosing bucket by NDV value of that bucket the
frequency of singleton bucket is calculated as:

singleton_freq = enclosing_bucket_freq / enclosing_bucket_width

The width value is in its turn computed as right bound minus left one
for LINT and DOUBLE representations of attribute type. Such computation
produces incredible frequencies for singleton buckets that built for
DATE type attribute. Because of dates are mappable to DOUBLE number of
microseconds from 2000-01-01 as it's done for other time related types
the width of any bucket for DATE type attribute becomes very huge and as
consequence frequency of signleton buckets produced from that bucket
steps down to zero. For 'equality with constant' predicates this leads
to zero selectivity and hence dummy cardinality for plan node to which
this predicate belongs.

The current patch replaces mapping for date values from DOUBLE number of
microseconds to LINT number of days that is a natural representation of
date value in PostgreSQL. Such modification doesn't allow to straightly
compare dates with other time related typed values in the context of
statistics estimation routines. But GPORCA in its current version
doesn't consider such comparisons and assigns to such predicates dummy
frequency value.