Added missing developer GUCs and commented out macros
There are some developer GUCs which are not represented in both unsync_guc_name.h and sync_guc_name.h:
1. log_btree_build_stats
2. optimize_bounded_sort
3. trace_syncscan
4. wal_debug

As a consequence, GPDB fails if the BTREE_BUILD_STATS, DEBUG_BOUNDED_SORT, TRACE_SYNCSCAN or WAL_DEBUG macros was defined when it was compiled. Also the first two macro definitions are not represented in pg_config_manual.h. The current patch eliminates this defect.