GPORCA: map DATE type values to LINT # of days instead of DOUBLE mcsecs
The commit 9b42761 had intoduced the change in computing of frequency
for singleton bucket produced from some enclosing one. Instead of
division frequency of enclosing bucket by NDV value of that bucket the
frequency of singleton bucket is calculated as:

singleton_freq = enclosing_bucket_freq / enclosing_bucket_width

The width is computed as right bound minus left one for both LINT and
DOUBLE representations of attribute type. Such computation produces a
very small frequency for singleton bucket that is built for DATE type
attribute. This is due to the fact that dates are mappable to DOUBLE
numbers of *microseconds* from 2000-01-01 and the width of any non-point
bucket becomes very huge while frequency of derived signleton bucket
tends to zero. For 'equality with constant' predicates this leads to
zero selectivity and hence to dummy cardinality for plan node with this
predicate.

The current patch replaces mapping for date values from DOUBLE number of
microseconds to LINT number of days that is a natural representation of
date value in PostgreSQL. Such modification doesn't allow to straightly
compare dates with other time related typed values in the context of
statistics estimation routines. But GPORCA in its current version
doesn't consider such comparisons and assigns to such predicates dummy
frequency value.

Backport from commit 1a68363