WIP: Mirrors restore
This implementation of mirrors' restore does not work. The mirrors are
restored successfully by gprecoverseg; but after a few seconds (~5s)
they are marked "down" in gp_segment_configuration, probably by FTS. The
reason for this is not quite clear at the moment. The mirror processes
continue to run normally after this happens.