Proof-of-concept: Mirror recovery
Implement mirror recovery for GPDB PITR test. In order to do that, make
a copy of postgresql.conf & postgresql.auto.conf files and restore them
after restart. This is necessary for a promoted GPDB cluster, as
otherwise walsender process does not start (and there is no
communication between primaries and mirrors).