Parallel consumption of WALs from primary & mirror. Refactoring
Implement a proof-of-concept of parallel consumption of WALs from
primary and mirror at once:
* Two pg_walreceive processes are run simultaneously, consuming WALs
from both primary and mirror, and putting them into a SINGLE directory.
* Last WAL file (.partial) still requires renaming

CHECKPOINT and restore point creation on mirror acting as primary is
tested. This does not seem to affect the recovery process.

The code is refactored and simplified