Demonstrate improper recovery
Demonstrate recovery does not work properly with GPDB in specific case.

When primary fails, and mirror is promoted to take its position, it
generates a .history file, which the recovery process can use to switch
to the newest (originating from the mirror) timeline. However, history
files are not reread before a WAL segment file is finished.

This scenario is demonstrated by 'faulty_timeline.sh'. It hides a
.history file in order to emulate the situation described above.

The resulting state of a database is the last consistent state obtained
on *primary* before an unfinished transaction (primary requires
replication for transactions to fininsh). The replica is then promoted
and is left in a consistent, but outdated state.