Detect client disconnection while running query and interrupt execution
Fixes the case when the server is executing a lengthy query and the
client breaks the connection. The operating system will be aware that
the connection is no more, but postgres node doesn't notice this,
because it doesn't try to read from or write to the socket while running
query. So we'll get a zombie connection. In theory, the query could be
one that runs for a million years, continues to chew up CPU and I/O and
occupies a connection slot that's sad. For this reason, the query have
to be interrupted as much earlier as possible.

Provides a new GUC check_client_connection_interval that can be used to
check whether the client connection has gone away, while running very
long queries. It is disabled by default.

For now this uses a non-standard Linux extension (also adopted by at
least one other OS).  POLLRDHUP is not defined by POSIX, and other OSes
don't have a reliable way to know if a connection was closed without
actually trying to read or write.

Backport from PostgreSQL commits c30f54ad7 and 22f6f2c1c .