Check if index pathkeys list is empty in find_usable_indexes() function
index_pathkeys might be NULL when target index does not provide
a desired sort order in query or when useful for sorting index
corresponding to inherited or child relation does not participate
in join.

Due to that the following fragment fails:
    CREATE TABLE clstr_tst (a int, b INT) DISTRIBUTED BY (a);
    CREATE INDEX clstr_tst_b_c ON clstr_tst (b);
    CREATE TABLE clstr_tst_inh () INHERITS (clstr_tst);
    SELECT a FROM clstr_tst ORDER BY 1

The solution is to check if index_pathkeys list is empty before using it.