Don't choose indexscan when you need a motion for a subplan (#172)
When you have subquery under a SUBLINK that might get pulled up, you should
not allow indexscans to be chosen for the relation which is the
rangetable for the subquery. If that relation is distributed and the
subquery is pulled up, you will need to redistribute or broadcast that
relation and materialize it on the segments, and cdbparallelize will not
add a motion and materialize an indexscan, so you cannot use indexscan
in these cases.
You can't materialize an indexscan because it will materialize only one
tuple at a time and when you compare that to the param you get from the
relation on the segments, you can get wrong results.

Because we don't pick indexscan very often, we don't see this issue very
often. You need a subquery referring to a distributed table in a subplan
which, during planning, gets pulled up and then when adding paths, the
indexscan is cheapest.

Backport @cd055f9