ORCA: Introduce two versions of CStackDescriptor::BackTrace
Implement an additional version of `CStackDescriptor::BackTrace()`. The
new implementation is to be used when x86 assembly instructions are not
available, and pure `backtrace()` is used instead. In this case, frames
are traversed in a different fashion, and there is no need in complex
return address calculation algorithm (relying on x86 implementation
details).