Remove platform specific stack unwinding
ORCA and GPDB relied on a platform specific stack unwinding that
required -fno-omit-frame-pointer during compilation. Though this
approach unwinds the stack faster than a backtrace() call it reserves
a register that is a bad practice. Also it is not an easy thing
to port ORCA to other platforms (we need to write assembly specific
code for that). Besides, gp_backtrace()'s fallback to backtrace()
was not sigserv safe for the builds without -fno-omit-frame-pointer.
So, this commit makes GPDB and ORCA use backtrace() instead of frame
pointers.