Convert float8 with double instead of long double
It seems that GPDB faced a glibc bug on RedHat 6 with denormalized
numbers and fixed it using two-step conversion in `float8in()`:

- text to long double
- long double to double after denormalized numbers validation

This differs from PG approach where we rely on libc and make the
direct conversion to double. RedHat 6 seems to be an old system and
there is no much sense to differ from upstream and have additional
performance penalty converting long double to double in a two-step
manner and making additional checks.

There is another problem with this code - long double can have
problems on some architectures that are currently not supported by
GPDB, but where it can be ported in the future. Originally this
problem was detected as research about porting GPDB to PowerPC:
it doesn't have analogue of Intel 80-bit long double with 63-bit
mantissa and extended range between 1.0E-4932W and 1.0E+4932W.
Instead, PowerPC has an extended double type combined from two
doubles with 106-bit mantissa and the same range as double provides
1.0E−323L - 1.0E+308L. As a result, we can't use the same validation
logic for the numbers near +/- infinity as it is possible to do on
Intel. So this GPDB specific code doesn't help us to solve
the problem on RedHat 6 on PowerPC and fails the float test.

So, let's get rid of it and return the code to PG approach with
a double.