pg_resgroup_move_query improvements.
The pg_resgroup_move_query function implemented in 51ee26be3c5562e73f1311b44feb3f0e31f4c1aa has several disadvantages.
1) Slot leaking. The slot acquired in caller process doesn't free if target process isn't alive, didn't receive a signal, or even received, but at the time it was in idle state.
2) Race condition between UnassignResGroup called at the end of transaction and handler called immediately after USR1 signal.

Old slots logic was tricky because of inter-process communication and necessity of acquiring slot in one process and freeing it in another. The solid architecture solution is to do this in one process, do not acquire a slot for a long without a need, and free it as soon as something goes wrong. To avoid slot waiting on already running process, the slot acquiring logic was changed - target process will produce a warning, but will keep the execution. The moving process was already async, but became more async than before with this patch.

To avoid race condition between UnassignResGroup and handler, handler is now executed based on the interruption mechanics, which caused handler to be called at the safe places not interrupting some random parts of code.

There are some more additions, like new exceptions, regression tests and docs changes.