auto_explain: fix failure when executor hooks are run in QEs (#358)
The reason of failure is that auto_explain hooks are fully executed on QEs.
The global variable Gp_role is initialized after auto_explain is loaded as shared preload library.
The QE has valid planstate objects just for nodes in query plan corresponded to the slice of current QE.
As result, for "intermediate" slices we have NULL pointer to subtree state for "child" (within slice borders)
Motion nodes (subtree itself is valid and points to real node but state for this subtree is not).
ExplainNode() routine called within ExplainPrintPlan() function fails trying to dereference that pointer.

To fix the failure I've changed the guard check Gp_role != GP_ROLE_DISPATCH in _PG_init() of auto_explain
to macro IS_QUERY_DISPATCHER() that relies on GpIdentity.segindex value to define the kind of node on which
code is executed. As a consequence, hooks of auto_explain are activated just on coordinator node.
To exclude execution of hooks on coordinator on entrydb kind of worker (having gp_role=executor)
I've added additional check Gp_role != GP_ROLE_DISPATCH into macro auto_explain_enabled().
This check disables auto_explain when some work (as executor) is doing at master.