Fix rows number count logic
When we ask GPDB to explain a query, it uses flow to count rows number,
and it can be either a single segment command, acommand can be
executed on all segments or a limited number of segments in general
cases. But there are cases when the flow is not generated, so GPDB
uses segments count by default, what can be incorrect for some nodes,
which are executed on a single segment only. This patch doesn't rely
on flow, it uses gangSize of a current slice which is calculated
correctly and always presents. gangSize is a simple and reliable way to
get a number of segments which processed some part of a plan.

Co-authored-by: Polina Bungina <bungina@gmail.com>