Early free underlying Scan structures
The tables with high count of partitions can drain high amount of executor
memory for processing. This can be especially critical for column-oriented
tables, because we allocating a buffer (which is triple of block size) for each
column. Following the standard flow we free such buffers at executor end. But it
seems, we can safely free some structures, including this buffers, after
processing all slots in `*getnextslot()` access method.

Related to #13725