orca: deduplicate target(result) relations
Duplication of target (result) relation is a problem, when a
final planned_statement, contains duplicates - final plan will
have got an incorrect `scanrelid's` at `SeqScan`, so epq mechanism
won't work correctly, because it relies on scanrelid, here is an
example: function `ExecScanFetch`:
```if (estate->es_epqTupleSet[scanrelid - 1])```

  To prevent duplcicates and correctly set indexes of scanrelid
at `planned_stmt->rtable` (which will be propagated to the segments),
here is used marking of target (result) relations. If current relation
processed at `CTranslatorQueryToDXL` is a result relation store info
at `CDXLTableDescr` about to which query it belongs (id > 0), if it's
not a result relation zero value is used. Also this info propagates
throgh other transformations and finally procceed at
`CTranslatorDXLToPlStmt`.
  At `CTranslatorDXLToPlStmt` each table descriptor validates, is it a
target (result) relation, if so and it was proceed, don't increment
index and don't add this table descr.