Remove per byte precision message truncation for multibyte strings
gpdb does not truncate error log correctly on error in external web tables.

If error code is not zero and error message is truncated in interpretError
on multi-byte character, then assert occurs.

If error code is zero and command emit stderr and this error message is
truncated in write_log (which is called by read_err_msg) on multi-byte
character, then bad character writes to log and later this log can not be
correctly read.

Solution is to correct error log truncation according to database encoding.