Fix "Failed to kill processes for segment" error of gpstop
Geenplum python scrips try to parse system command’s STDOUT and STDERR
in English and may fail if locale is different from en_US. This patch
forced it to check the return code rather then a message itself.

Also, this patch reinstalls glibc-common in docker container. This is
necessary to get langpacks in docker because docker images don't contain
them.