ADBDEV-3723: Flow isn't set properly for Shared Scan nodes during translation to physical plan
ORCA does not always correctly fill in the flow for shared scans,
although the shareinput_mutator_xslice_* functions in cdbmutate
require a correctly filled flow.
These functions, among other things, check that if at least one part
of a shared scan is executed on the coordinator, then all its other parts
must also be performed on the coordinator.
They do this on the shared scan flow, assuming that if
flotype = FLOW_SINGLETON and segindex < 0, then
the shared scan will be performed on the coordinator.
For such a check, flow must be filled in for a shared scan.
ORCA always fills the flow in the shared scan,
but sometimes it does this incorrectly.
ORCA correctly fills the flow only under the Motion nodes, and if
there is no Motion above the producers or consumers of the shared scan,
then the flow of this shared scan will be filled incorrectly
with the parameters flotype = FLOW_UNDEFINED and segindex = 0,
which is not true if the slice will be executed only on
one of the segments or even on the coordinator.
The executor determines where the slice will be executed using gangType in
the FillSliceGangInfo function, and flotype and segindex are used
to determine gangType only under Motion nodes.
This shows that the flow of a shared scan has no effect on
where the slice will be executed, as for this, the flow of the node
under the Motion of the slice in which the shared scan is located is used.
To determine where the slice will be executed, we can use the node's flow
under Motion, keeping and checking that the slice is being executed on
the coordinator by the conditions flotype = FLOW_SINGLETON and segindex < 0.
We get rid of the need to fill in the flow for a shared scan in ORCA,
and now we can remove the function that did this.