arenadata_toolkit: initial implementation of extension
This patch provides initial version of arenadata_toolkit extension to
simplify versioning of arenadata_toolkit objects (and work with them)
at adb-bundle. There are implemented two extension scripts:
1. arenadata_toolkit--unpackaged--1.0 - used to create extension if
   toolkit was manually installed at bundle (updating existing
   installation)
2. arenadata_toolkit--1.0 - used to create extension if there is no
   toolkit at bundle (new installation)

Behavour of new installation and updating differs: the table
db_files_history in case of updating (1) binds to the extension, while
a new installation of toolkit (2) won't create (in turn won't bind)
this table, because of problems with AO-segment tables in
`ALTER EXTENSION ADD/DROP TABLE` (ADBDEV-3738). Thus creating and
binding this table to the extension is a bundle responsibility.