Vacuum make all segments inaccessible in case of concurrent transaction present
Vacuum analyzes necessity of compaction segment files of append-optimized
table only on query executors. And after compaction it always marks
(in the hashmap on the coordinator) the segment files as awaiting drop,
regardless of their actual state on segments. But drop phase may be skipped
in case of there are concurrent transactions and segment files are left
in awaiting drop state (in the hashmap on the coordinator). While
(on the segments) the actual compaction might not have happened, and there
the segment files might have remained in their normal state. On intensive
concurrent insertion into the table, the vacuum will mark (in the hashmap
on the coordinator) all available segment files as awaiting drop and
the next insert will fail.

ERROR: could not find segment file to use for inserting into relation

The patch extracts information from the segments about the actual state
of the segment files after compaction and marks (in the hashmap on
the coordinator) as awaiting drop only those segment files that
were actually compacted. This allows for intensive concurrent insertion
into the table.