ORCA produces bogus plan for queries with CTE during handling distribution for Sequence children
The condition for checking the requested distribution does not work
in case of inlining the outer CTE by a filter, because the filter replaces
the requested distribution.

Therefore, for cases of CTE inlining by a filter, I added a flag that remembers
this, and then if this flag is set, then the filter passes the requested
distribution, and does not replace it.