Don't push upper predicates down to modifying CTE
For queries, which contain DML operations inside the WITH clause,
pushing down restrictions from upper query levels can lead to conditional table
modification or cause planner errors. That behaviour is undesirable, because in
context of modifying CTEs it is expected, that the DML operation will affect the
whole table, and the restriction at upper level is applied to the result of
RETURNING clause.
Therefore, this commit proposes limitations that don't allow planner to
push the predicates down from upper level to CTEs, whose commandType is
non-SELECT. The if-clause inside set_cte_pathlist function was extended
with additional condition, which checks the commandType of CTE's subquery.

Co-Authored-By: Alexey Gordeev <goa@arenadata.io>