ORCA initialization refactoring
When initializing CGPOptimizer::InitGPOPT in the gpos_init function,
ORCA mixes the C approach (returning error codes and their further checking)
and the C++ approach (thrown exceptions and their further interception).

1) If there is not enough memory or any other exception, the
CWorkerPoolManager::Init() function will not return GPOS_OK, and then the
gpos_init function will call the
CMemoryPoolManager::GetMemoryPoolMgr()->Shutdown() function, which will
nullify CMemoryPoolManager::m_memory_pool_mgr.

In this case, all the exceptions that have occurred will not be reraised,
which means that the execution will continue, and the next function
gpdxl_init will be called, in which there is an initialization of the
CAutoMemoryPool class.

In the constructor of this class, the static method
CMemoryPoolManager::GetMemoryPoolMgr() is called, which will return NULL,
because up the stack, we zeroed out m_memory_pool_mgr. Thus, a segfault
can occur when there is not enough memory.

2) If during the initialization of the ORCA an exception occurs that is
not caught by the initialization functions, then it is caught in the
InitGPOPT() function and can be silently ignored by simply disabling the ORCA.

Further, when the process terminates (even successfully), the
gpopt_terminate() function is called.

It tries to release the memory pool on the debug build
CMemoryPoolManager::GetMemoryPoolMgr()->Destroy(mp).

But due to an exception during the initialization of the ORCA,
this memory pool was not created and therefore the assert
GPOS_ASSERT(NULL != mp) is triggered.

On the debug build, the assert turns into an exception, which is
safely caught and muted in the TerminateGPOPT() function.

At the same time, the Shutdown method of the CWorkerPoolManager class
is not called, in which the CWorkerPoolManager::m_worker_pool_manager
variable is set to zero.

When deleting the MemoryContextDelete(OptimizerMemoryContext) context on
the debug version, all freed memory is filled with the sequence 7x,
including the m_single_worker variable of the m_worker_pool_manager
instance of the CWorkerPoolManager class.

At the very end of the process, the destructor
CGroupExpression::~CGroupExpression is called on the global
static variable CGroupExpression::m_gexprInvalid.

And then the destructor of its parent class CRefCount::~CRefCount, in which
(again, on a debug build) a sequence of methods is called
ITask::Self()
IWorker *worker = IWorker::Self()
worker = CWorkerPoolManager::WorkerPoolManager()->Self()
which results in worker = worker = 0x7f7f7f7f7f7f7f7f and a
segfault occurs on return worker->GetTask().

I refactored the ORCA initialization code, getting rid of the C approach,
i.e. removed the return of error codes and their further checking,
added C++ approach, i.е. interception of exceptions and their further
re-raising, and also translated the shutdown, cleanup and destroy
methods into static ones.